Feat/optimize process service (#247)
* split process_step

* optimize

* fix

* optimizeCode
